Refactor minimum_table.py
